\docType{package}
\name{DySeq}
\alias{DySeq}
\alias{DySeq-package}
\title{DySeq Overview}
\description{A collection of functions for dyadic sequence anaylses.}
\details{
  DySeq provides an implementation of Bakeman & Gottman's (1997) approach of aggregated logit models, see help(LogSeq). And some additional functions for data preparation. 
}
\author{Peter Fuchs}
\keyword{package}
