% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstFreq.R
\name{EstFreq}
\alias{EstFreq}
\title{EstFreq}
\usage{
EstFreq(x, t, min.cell = 5, k = 20000)
}
\arguments{
\item{x}{a matrix containing the assumed probabilities for the expected transition tables}

\item{t}{number of time intervalls}

\item{min.cell}{a single integer defines what counts as a low frequency (5 by convention)}

\item{k}{Number of simulations (at least 20.000 is recommended)}
}
\description{
Simulate number low freq or zero frequencies for a state-transition table dependend ont the number of time points
}
\details{
First for entrys represent the first column, fith to eights entry the second column
}
\examples{
my.trans.table<-matrix(c(0.57, 0.13,0.05,0.05,0.05, 0.05,0.05,0.05),4,2)

my.cellproblems<-EstFreq(my.trans.table, 100, 5, k=20000)

my.cellproblems


}

