% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstTime.R
\name{EstTime}
\alias{EstTime}
\title{EstTime}
\usage{
EstTime(x, t = NA, crit = 0.05, zero = T, min.cell = 5, k = 20000,
  pos = "bottomleft", smoothed = T, show.it = 10, max.it = 10000)
}
\arguments{
\item{x}{a matrix containing the assumed probabilities for the expected transition tables}

\item{t}{optinol: a vector of time intervalls for which frequencies should be simulated}

\item{crit}{optional, but must be sepcified is t is not: Simulations will end if zero frequencies are less than crit}

\item{zero}{optional: if FALSE Simulations will end if number of low frequency problems is less than crit}

\item{min.cell}{a single integer defines what counts as a low frequency (5 by convention)}

\item{k}{Number of simulations (at least 20.000 is recommended)}

\item{pos}{position of the outputs legend. Options are: "bottomleft", "bottomright", "upperleft", and "upperright".}

\item{smoothed}{logical value. If true, outputlines will be smoothed!}

\item{show.it}{single integer that defines which steps of iteration protocol should be shown. Only active if t is not defined else iterationprotocoll is replaced with a progression bar}

\item{max.it}{single integer that defines the maximum number of iterations if t is not specified.}
}
\value{
a EstTime object; a list of three vectors that will, if printed, provide a plot of expected number of low and zero cell frequencies
}
\description{
Plots simulated number low freq or zero frequencies for a state-transition table against the number of time points
}
\details{
First vector represents time points, second frequency of cases with zero's, thrid frequency of cases with low cell frequencies
}
\examples{


}

