% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GiveSome}
\alias{GiveSome}
\title{Dyadic sequences of 45 subjects engaging a social dilemma}
\format{An object of class \code{matrix} with 42 rows and 62 columns.}
\source{
Halstenberg, E. (2016). The effect of social value orientation on cooperation in a four-coin dilemma: a quasi-replication study using the svo slider measure. (Unpublished bachelor thesis). University Bielefeld, Germany.
}
\usage{
GiveSome
}
\description{
The data set stems from the bachelor thesis of Halstenberg (2016) and
contains sequences of 45 subjects that engaged in a 32-rounds-long 
four-coin dilemma. That is, each player starts with four coins that 
are worth one point for oneself and two points for the opponent. Both
players have to submit zero to four of them to the other player. The 
decision is made secretly and simultaneously.
}
\details{
The computer was set to ignore the humans behavior at all. Instead the 
32 rounds were divided into eight blocks. Within each block, the computer
gave one-times one, two-times two, and one-times three coins in randomized
order. The only exception was the very first turn, in which the algorithm
always gave two coins followed by one, two and three coins in randomized 
order.Thus, on average, the algorithm gave two coins. Hence, it was 
always possible for the human player to give more or fewer coins than the
algorithm did before.

For the humnan, it was coded whether the human player gave more (1) coins 
in his turn than the algorithm did in the last turn or not (0). The same 
was coded for the computer. Thus, coding started in the second turn 
resulting in 31 entries for each of both. 
 
The data frame contains 45 rows (subjects) and 62 columns. Columns 1 to
31 correspond to the human behavior, columns 32:61 to the algorithm.
}
\keyword{datasets}
