% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_Trans.R
\name{ML_Trans}
\alias{ML_Trans}
\title{ML_Trans}
\usage{
ML_Trans(data, first, second)
}
\arguments{
\item{data}{a data.frame, which contains dyadic sequences in a bride format}

\item{first}{column-index that defines the first sequence of each dyad}

\item{second}{column-index that defines the second sequence of each dyad}
}
\description{
Transforms transition tables into multi-level data. 
Each transition between states is handled as a single 
observation, which is nested within dyads.
}
\examples{

# Example: Sequences from couples cope into multi-level data

data(CouplesCope)
ML_data<-ML_Trans(CouplesCope, 2:49, 50:97)
}

