% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonCumHaz.R
\name{NonCumHaz}
\alias{NonCumHaz}
\title{NonCumHaz}
\usage{
NonCumHaz(x, t = NA, plot = FALSE)
}
\arguments{
\item{x}{a vector containing cumulated hazard or a survfit-object from the 'survival'-package}

\item{t}{optional vectorcontainting time reference for x (is required for plot)}

\item{plot}{logical value indicating if non-cumulated plot should be generated}
}
\description{
Computes the non-cumulated hazard from cumulated hazard
}
\examples{

# Example 1: Short artificial data
# example cumulated hazard with time referenz
cumhaz<-c(0.2 ,0.21 ,0.31 ,0.44 ,1.1  ,1.1  ,1.12 ,1.2)
time<-c(4     ,5    ,6    ,7    ,10   ,14   ,15   ,16)

NonCumHaz(cumhaz, time, plot=T)



# Example 2: Every hazard entry represents one point of time

# if every hazard entry repesents one point of time
NonCumHaz(cumhaz, 1:8, plot=T)



# Example 3: real data and real researchquestion
# install.packages("survival")
library(survival)

# How long till the last stress signal
my.last<-LastOccur(CouplesCope[,50:97],1)

# If last stress signal was in time intervall 48,
# stress did not end till the observation duration
event<-rep(1,length(my.last))
event[my.last==48]<-0
my.surv<-Surv(my.last,event) # creates a object for survival time analysis
my.fit<-survfit(coxph(my.surv~1)) # fits coxregression without covariates on the data
plot(my.fit) # survival curve
plot(my.fit, fun="cumhaz") # cumulated survival curve
NonCumHaz(my.fit, plot=T)
NonCumHaz(my.fit$cumhaz, my.fit$time, plot=T) # if over packages than 'survival'are used
}

