% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.LogSeq.R
\name{plot.LogSeq}
\alias{plot.LogSeq}
\title{plot.LogSeq}
\usage{
\method{plot}{LogSeq}(x)
}
\arguments{
\item{x}{a LogSeq object, that should be printed}
}
\description{
Generates interaction diagram for LogSeq Objects
}
\details{
#'
}

