% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.EstFreq.R
\name{print.EstFreq}
\alias{print.EstFreq}
\title{print.state.trans}
\usage{
\method{print}{EstFreq}(x, ...)
}
\arguments{
\item{x}{a EstFrag object, that should be printed. See help(EstFreq)}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generates output for EstFrag object
}

