% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.EstTime.R
\name{print.EstTime}
\alias{print.EstTime}
\title{print.EstTime}
\usage{
\method{print}{EstTime}(x, pos = NA, ...)
}
\arguments{
\item{x}{a EstTime object, printing it will result in a frequency plot. See help(EstTime)}

\item{pos}{position of legend}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generates output for EstTime object
}

