% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.LogSeq.R
\name{print.LogSeq}
\alias{print.LogSeq}
\title{print.LogSeq}
\usage{
\method{print}{LogSeq}(x, ...)
}
\arguments{
\item{x}{a LogSeq object, that should be printed}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generates output for LogSeq Objects, see: \code{\link{LogSeq}}
}

