% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.state.trans.R
\name{print.state.trans}
\alias{print.state.trans}
\title{print.state.trans}
\usage{
\method{print}{state.trans}(x)
}
\arguments{
\item{x}{a state.trans object, that should be printed}
}
\description{
Generates output for state.trans object
}

