% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSeq.R
\name{simSeq}
\alias{simSeq}
\title{simSeq}
\usage{
simSeq(trans, initial, length)
}
\arguments{
\item{trans}{a 4x4 matrix containing transition probabilities}

\item{initial}{a four element vector containing initial states probabilities}

\item{length}{single value specifying the length of the simulated sequence}
}
\description{
Simulates a single sequence with four possible states.
}
\examples{

test1<-matrix(c(0.5 , 0.2 , 0.2 , 0.1,
                0.8 , 0.05, 0.05, 0.1,
                0.5 , 0.1 , 0.2 , 0.2,
                0.1 , 0.1 , 0.1 , 0.7) , 4 , 4 , byrow = TRUE)

initial<-c(.25 , .25 , .25 , .25)

simSeq(test1, initial, 30)

}

