% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSeqSample.R
\name{simSeqSample}
\alias{simSeqSample}
\title{simSeqSample}
\usage{
simSeqSample(trans, initial, length, N)
}
\arguments{
\item{trans}{a 4x4 matrix containing transition probabilities}

\item{initial}{a four element vector containing initial states probabilities}

\item{length}{single value specifying the length of the simulated sequence (columns)}

\item{N}{number of sequences which should be simulated (rows)}
}
\description{
Simulates N single sequences with four possible states.
}
\examples{

test1<-matrix(c(0.5 , 0.2 , 0.2 , 0.1,
                0.8 , 0.05, 0.05, 0.1,
                0.5 , 0.1 , 0.2 , 0.2,
                0.1 , 0.1 , 0.1 , 0.7) , 4 , 4 , byrow = TRUE)

initial<-c(.25 , .25 , .25 , .25)

simSeq(test1, initial, 30)

}
