% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.LogSeq.R
\name{single.LogSeq}
\alias{single.LogSeq}
\title{single.LogSeq}
\usage{
single.LogSeq(x, case)
}
\arguments{
\item{x}{a LogSeq object, that should be printed.}

\item{case}{determines which case should be shown.}
}
\description{
prints estimates vor single case out of an LogSeq object
}
\examples{
data(CouplesCope)
my.states<-StateExpand(CouplesCope, 2:49, 50:97)
my.trans<-StateTrans(my.states, FALSE)
my.logseq<-LogSeq(my.trans)
my.logseq
single.LogSeq(my.logseq, 41)

}

