% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateTrans.R
\name{StateTrans}
\alias{StateTrans}
\title{StateTrans}
\usage{
StateTrans(x, first = TRUE, dep.lab = c("1", "0"), indep.lab = c("1-1",
  "1-0", "0-1", "0-0"))
}
\arguments{
\item{x}{x Dataframe or matix containing combined sequences, see help(StateExpand)}

\item{first}{a logical value indicating if the first sequence should used as dependend variable (TRUE) or the second (FALSE)}

\item{dep.lab}{a two-element string vector with labels dependend variable (first level zero, second corresponds to level one)}

\item{indep.lab}{a four-element string vector with labels for the combined variable (order is correspondung to the order of the StateExpand function)}
}
\description{
Produces state transition tables for dyadic binary sequences.
For an extensive overvie see Kenny, Kashy and Cook (2006).
The original idea stems (as far as known) from Bakeman and Gottman (1997).
}
\examples{

# Example 1: Sequences from couples cope

data(CouplesCope)
my.s<-StateExpand(CouplesCope, 2:49, 50:97)
StateTrans(my.s) # First sequence is dependend variable (what behavior preceeds stress signals?)
StateTrans(my.s, FALSE) # Second sequence is dependend variable (what behavior preceeds dyadic coping signals?)
StateTrans(my.s, FALSE)[[41]] # investigating a single case

}
\references{
Bakeman, R., & Gottman, J. M. (1997). Observing interaction: An introduction to sequential analysis. Cambridge university press.

Kenny, D. A., Kashy, D. A., & Cook, W. L. (2006). Dyadic data analysis. Guilford Press.
}

