% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Markov_as_APIM.R
\name{Basic_Markov_as_APIM}
\alias{Basic_Markov_as_APIM}
\title{Basic_Markov_as_APIM}
\usage{
Basic_Markov_as_APIM(x, first, second, boot = 1000, SimOut = FALSE,
  CPU = 1, sim = "ordinary", parallel = "no")
}
\description{
Fits a basic Markov-model on dyadic sequences. 
The transition matrix is converted into equivalent APIM-beta-coefficients.
Bootstrapping is used for approximating p-values. (H1: Effect is different from zero)
}
\examples{
\dontrun{

trans1<-APIMtoTrans(B0_1=0, AE_1=1, PE_1=0, Int_1=0,
                   B0_2=0, AE_2=0, PE_2=0, Int_2=0)

x<-simSeqSample(trans=trans1, initial=rep(.25,4), length=100, N=100)

# small boot-size sample only for demonstration purposes! 
MasAPIM(x, 1:100, 101:200, boot=10)
}
}

