% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstFreq.R
\name{EstFreq}
\alias{EstFreq}
\title{EstFreq}
\usage{
EstFreq(x, t, min.cell = 5, k = 20000)
}
\arguments{
\item{x}{a matrix containing the assumed probabilities for the expected transition tables}

\item{t}{number of time intervalls}

\item{min.cell}{a single integer defines what counts as a low frequency (5 by convention)}

\item{k}{Number of simulations (at least 20.000 is recommended)}
}
\description{
Simulate number low freq or zero frequencies 
for a state-transition table dependend on
the number of time points
}
\details{
the matrix must have 2*4 dimensions

first column represents if behavior of interest is shown

Second column represents if behavior of interest is not shown

first row: behavior of interest of both partners was shown

second row: behavior of interest was shown only by the partner

third row: behavior of interest was shown only by the actor

second row: behavior of interest was not shown
}
\examples{
\dontrun{
my.trans.table<-matrix(c(0.57, 0.13,0.05,0.05,0.05, 0.05,0.05,0.05),4,2)

my.cellproblems<-EstFreq(my.trans.table, 100, 5, k=20000)

my.cellproblems
}

}

