% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.LogSeq.R
\name{plot.LogSeq}
\alias{plot.LogSeq}
\title{plot.LogSeq}
\usage{
\method{plot}{LogSeq}(x, y, ...)
}
\arguments{
\item{x}{a LogSeq object, that should be printed.}

\item{y}{further arguments passed to or from other methods.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generates interaction diagram for LogSeq Objects
}
\details{
#'
}

